integrations/langfuse/example/requirements.txt
langfuse-deepstack
datasets
sentence-transformers